Add the resettable option; split the transport in two to avoid forced ResetInterface implementation
